+++ test/errors/handle.test.ts
@@ -1,5 +1,5 @@
 import {expect, fancy} from 'fancy-test'
-import * as fs from 'fs-extra'
+import {readFileSync} from 'node:fs'
 import * as path from 'path'
 import * as process from 'process'
 
@@ -106,7 +106,7 @@ describe('handle', () => {
     handle(new CLIError('uh oh!'))
     expect(ctx.stderr).to.equal(` ${x}   Error: uh oh!\n`)
     await config.errorLogger!.flush()
-    expect(fs.readFileSync(errlog, 'utf8')).to.contain('Error: uh oh!')
+    expect(readFileSync(errlog, 'utf8')).to.contain('Error: uh oh!')
     expect(process.exitCode).to.equal(2)
   })